+++ libs/xui/src/button/button.component.ts
@@ -23,7 +23,7 @@ export class XuiButtonComponent {
 
   @Input() xType: 'normal' | 'dashed' | 'stroked' | 'raised' | 'fab' | 'icon' | string = 'normal';
   @Input() xSize: 'sm' | 'md' | 'lg' | string = 'md';
-  @Input() xColor: 'primary' | 'primary-alt' | 'secondary' | 'destructive' | 'neutral' | 'minimal' | string = 'primary';
+  @Input() xColor: 'primary' | 'primary-alt' | 'secondary' | 'error' | 'neutral' | 'minimal' | string = 'primary';
 
   @Input() @InputBoolean() disabled = false;
   @Input() xClick: () => Promise<boolean>;