+++ libs/xui/src/checkbox/checkbox.component.ts
@@ -31,7 +31,7 @@ export class XuiCheckboxComponent implements ControlValueAccessor, OnInit {
   onTouched = () => {};
 
   @Input() @InputBoolean() disabled = false;
-  @Input() color: 'success' | 'warn' | 'information' | 'primary' | 'destructive' | string = 'success';
+  @Input() color: 'success' | 'warning' | 'info' | 'primary' | 'error' | string = 'success';
 
   @Input()
   get value() {